sqlalchemy - Member CRUD

+++ sqlalchemy02.py
@@ -1,34 +1,38 @@
-from typing import List
 from datetime import datetime
+from typing import List, Optional
 
-from fastapi import FastAPI, Depends
+from fastapi import FastAPI
+from fastapi.params import Depends
 from pydantic import BaseModel
-from sqlalchemy import create_engine, Column, String, Integer, DateTime, func
+from sqlalchemy import create_engine, Column, String, Integer, Sequence, DateTime, func
 from sqlalchemy.orm import sessionmaker, declarative_base, Session
 
-sqlite_url = 'sqlite:///python.db'
-engine = create_engine(sqlite_url,
-                       connect_args={'check_same_thread': False}, echo=True)
-SessionLocal = sessionmaker(
-    autocommit=False, autoflush=False, bind=engine)
+from sqlalchemy01 import SessionLocal
 
-# 데이터베이스 모델 정의
+# 회원정보를 이용한 SQL CRUD
+# mno, userid, passwd, name, email, regdate
+
+sqlite_url = 'sqlite:///python.db'
+engine = create_engine(sqlite_url, connect_args={}, echo=True)
+SessionLocal = sessionmaker(autocommit=False,
+                            autoflush=False, bind=engine)
+
 Base = declarative_base()
 
 class Member(Base):
-    __tablename__ = 'members'
+    __tablename__ = 'member'
 
-    mno = Column(Integer, primary_key=True, index=True)
+    mno = Column(Integer, Sequence('seq_member'),
+                 primary_key=True, index=True)
     userid = Column(String, index=True)
     passwd = Column(String)
     name = Column(String)
     email = Column(String)
-    regdate = Column(DateTime(timezone=True), server_default=func.now())
+    regdate = Column(DateTime(timezone=True),
+                     server_default=func.now())
 
-# 데이터베이스 테이블 생성
 Base.metadata.create_all(bind=engine)
 
-# 데이터베이스 세션을 의존성으로 주입하기 위한 함수
 def get_db():
     db = SessionLocal()
     try:
@@ -36,60 +40,67 @@ def get_db():
     finally:
         db.close()
 
-# Pydantic 모델
-class MemberModel(BaseModel):
-    mno: int
+class NewMemberModel(BaseModel):
     userid: str
+    passwd: str
     name: str
     email: str
+
+class MemberModel(NewMemberModel):
+    mno: int
     regdate: datetime
 
-# FastAPI 애플리케이션 생성
+
 app = FastAPI()
 
 @app.get('/')
 def index():
-    return 'Hello, SQLAlchemy and FastAPI!'
+    return 'Hello, sqlalchemy!!, again'
 
-# 회원정보 조회
-@app.get('/members', response_model=List[MemberModel])
-def read_members(db: Session = Depends(get_db)):
+
+@app.get('/member', response_model=List[MemberModel])
+def read_member(db: Session = Depends(get_db)):
     members = db.query(Member).all()
     return members
 
-# 회원정보 추가
-@app.post('/members', response_model=MemberModel)
-def meadd(member: MemberModel, db: Session = Depends(get_db)):
-    db_member = Member(**member.dict())
-    db.add(db_member)
-    db.commit()
-    db.refresh(db_member)
-    return db_member
 
-# 회원정보 상세 조회
-@app.get('/members/{mno}', response_model=MemberModel)
-def read_member(mno: int, db: Session = Depends(get_db)):
+# @app.post('/member', response_model=NewMemberModel)
+@app.post('/member', response_model=str)
+def add_member(m: NewMemberModel, db: Session = Depends(get_db)):
+    m = Member(**dict(m))
+    db.add(m)
+    db.commit()
+    db.refresh(m)
+    # return m
+    return '데이터 입력 성공!!'
+
+
+@app.get('/member/{mno}', response_model=Optional[MemberModel])
+def readone_member(mno: int, db: Session = Depends(get_db)):
     member = db.query(Member).filter(Member.mno == mno).first()
     return member
 
-# 회원정보 삭제
-@app.delete('/members/{mno}', response_model=MemberModel)
+
+@app.delete('/member/{mno}', response_model=Optional[MemberModel])
 def delete_member(mno: int, db: Session = Depends(get_db)):
     member = db.query(Member).filter(Member.mno == mno).first()
-    db.delete(member)
-    db.commit()
+    if member:
+        db.delete(member)
+        db.commit()
     return member
 
-# 회원정보 수정
-@app.put('/members/{mno}', response_model=MemberModel)
-def update_member(mno: int, db: Session = Depends(get_db)):
-    member = db.query(Member).filter(Member.mno == mno).first()
-    for key, value in dict(exclude_unset=True).items():
-        setattr(member, key, value)
-    db.commit()
-    db.refresh(member)
+
+@app.put('/member/{mno}', response_model=Optional[MemberModel])
+def update_member(m: MemberModel, db: Session = Depends(get_db)):
+    member = db.query(Member).filter(Member.mno == m.mno).first()
+    if member:
+        for k, v in m.dict().items():
+            setattr(member, k, v)
+        db.commit()
+        db.refresh(member)
     return member
 
-if __name__ == "__main__":
+
+if __name__ == '__main__':
     import uvicorn
     uvicorn.run('sqlalchemy02:app', reload=True)
